refactoring: extraced workActivity into MainFragment
