+++ app/src/main/java/seemoo/fitbit/tasks/InformationTask.java
@@ -1,6 +1,7 @@
 package seemoo.fitbit.tasks;
 
 
+import seemoo.fitbit.activities.MainFragment;
 import seemoo.fitbit.activities.WorkActivity;
 
 /**
@@ -9,17 +10,17 @@ import seemoo.fitbit.activities.WorkActivity;
 class InformationTask extends Task {
 
     private Tasks tasks;
-    private WorkActivity activity;
+    private MainFragment mainFragment;
 
     /**
      * Creates an instance of information task.
      *
      * @param tasks    The instance of tasks.
-     * @param activity The current activity.
+     * @param mainFragment The current mainFragment.
      */
-    InformationTask(Tasks tasks, WorkActivity activity) {
+    InformationTask(Tasks tasks, MainFragment mainFragment) {
         this.tasks = tasks;
-        this.activity = activity;
+        this.mainFragment = mainFragment;
     }
 
     /**
@@ -28,11 +29,11 @@ class InformationTask extends Task {
      */
     @Override
     public void execute() {
-        activity.runOnUiThread(new Runnable() {
+        mainFragment.getActivity().runOnUiThread(new Runnable() {
 
             @Override
             public void run() {
-                activity.collectBasicInformation();
+                mainFragment.collectBasicInformation();
             }
         });
         tasks.taskFinished();
