+++ app/src/main/java/seemoo/fitbit/tasks/Tasks.java
@@ -3,6 +3,7 @@ package seemoo.fitbit.tasks;
 
 import android.bluetooth.BluetoothDevice;
 
+import seemoo.fitbit.activities.MainFragment;
 import seemoo.fitbit.activities.WorkActivity;
 import seemoo.fitbit.https.HttpsClient;
 import seemoo.fitbit.interactions.Interactions;
@@ -17,20 +18,19 @@ public class Tasks {
     private final String TAG = this.getClass().getSimpleName();
 
     private Interactions interactions;
-    private WorkActivity activity;
+    private MainFragment mainFragment;
     private TaskQueue mTaskQueue;
 
     /**
      * Creates an instance of tasks.
      *
      * @param interactions  The instance of interactions
-     * @param activity      The current activity.
-     * @param buttonHandler The isntance of the button handler.
+     * @param mainFragment      The current mainFragment.
      */
-    public Tasks(Interactions interactions, WorkActivity activity, ButtonHandler buttonHandler) {
+    public Tasks(Interactions interactions, MainFragment mainFragment) {
         this.interactions = interactions;
-        this.activity = activity;
-        mTaskQueue = new TaskQueue(buttonHandler);
+        this.mainFragment = mainFragment;
+        mTaskQueue = new TaskQueue();
     }
 
     /**
@@ -78,12 +78,12 @@ public class Tasks {
      */
     public void taskUploadDump(HttpsClient client, BluetoothDevice device, String type) {
         if (!interactions.getAuthenticated()) {
-            mTaskQueue.addTask(new InteractionsTask(interactions, "authentication", this, activity));
+            mTaskQueue.addTask(new InteractionsTask(interactions, "authentication", this, (WorkActivity) mainFragment.getActivity()));
         }
-        if (activity.getDataFromInformation(type) == null || activity.wasInformationListAlreadyUploaded(type)) {
-            mTaskQueue.addTask(new InteractionsTask(interactions, type, this, activity));
+        if (mainFragment.getDataFromInformation(type) == null || mainFragment.wasInformationListAlreadyUploaded(type)) {
+            mTaskQueue.addTask(new InteractionsTask(interactions, type, this, (WorkActivity) mainFragment.getActivity()));
         }
-        mTaskQueue.addTask(new UploadDumpTask(client, device, activity, type, this));
+        mTaskQueue.addTask(new UploadDumpTask(client, device, mainFragment, type, this));
         if (type != ConstantValues.INFORMATION_MICRODUMP) {
             mTaskQueue.addTask(new InteractionsTask(interactions, type + "Upload", this, client));
         }
@@ -100,7 +100,7 @@ public class Tasks {
     public void taskStartup(Interactions interactions, WorkActivity activity) {
         mTaskQueue.addTask(new InteractionsTask(interactions, ConstantValues.INFORMATION_MICRODUMP, this, activity));
         mTaskQueue.addTask(new LoadSettingsTask(this, activity));
-        mTaskQueue.addTask(new InformationTask(this, activity));
+        mTaskQueue.addTask(new InformationTask(this, mainFragment));
         mTaskQueue.addTask(new EmptyTask(this));
     }
 }
