+++ app/src/main/java/seemoo/fitbit/tasks/UploadDumpTask.java
@@ -2,6 +2,7 @@ package seemoo.fitbit.tasks;
 
 import android.bluetooth.BluetoothDevice;
 
+import seemoo.fitbit.activities.MainFragment;
 import seemoo.fitbit.activities.WorkActivity;
 import seemoo.fitbit.https.HttpsClient;
 import seemoo.fitbit.miscellaneous.ConstantValues;
@@ -13,7 +14,7 @@ class UploadDumpTask extends Task {
 
     private HttpsClient client;
     private BluetoothDevice device;
-    private WorkActivity activity;
+    private MainFragment mainFragment;
     private String type;
     private Tasks tasks;
 
@@ -21,14 +22,14 @@ class UploadDumpTask extends Task {
      * Creates an instance of upload dump task.
      * @param client The instance of https client.
      * @param device The device where the dump comes from.
-     * @param activity The current activity.
+     * @param mainFragment The current activity.
      * @param type The type of the dump.
      * @param tasks The instance of tasks.
      */
-    UploadDumpTask(HttpsClient client, BluetoothDevice device, WorkActivity activity, String type, Tasks tasks){
+    UploadDumpTask(HttpsClient client, BluetoothDevice device, MainFragment mainFragment, String type, Tasks tasks){
         this.client = client;
         this.device = device;
-        this.activity = activity;
+        this.mainFragment = mainFragment;
         this.type = type;
         this.tasks = tasks;
     }
@@ -39,11 +40,11 @@ class UploadDumpTask extends Task {
      */
     @Override
     public void execute() {
-        String data = activity.getDataFromInformation(type);
+        String data = mainFragment.getDataFromInformation(type);
         if(data != null) {
             int cutOff = data.indexOf(ConstantValues.RAW_OUTPUT);
             client.uploadDump(data.substring(cutOff + ConstantValues.RAW_OUTPUT.length()), device.getName(), tasks);
-            activity.setInformationListAsAlreadyUploaded(type);
+            mainFragment.setInformationListAsAlreadyUploaded(type);
         } else{
             tasks.taskFinished();
         }
