+++ app/src/main/res/layout/activity_work.xml
@@ -7,15 +7,12 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true">
 
-    <android.support.constraint.ConstraintLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
+    <LinearLayout
+        android:orientation="vertical"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context="seemoo.fitbit.activities.WorkActivity">
+        android:layout_height="match_parent">
 
-        <android.support.v7.widget.Toolbar
+    <android.support.v7.widget.Toolbar
             android:id="@+id/work_activity_toolbar"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -24,115 +21,13 @@
             app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
             />
 
-        <android.support.design.widget.FloatingActionButton
-            android:id="@+id/button_WorkActivity_7"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="200dp"
-            android:layout_marginRight="8dp"
-            android:layout_x="287dp"
-            android:layout_y="140dp"
-            android:clickable="true"
-            android:onClick="clearAlarmsButton"
-            app:backgroundTint="?attr/colorButtonNormal"
-            app:fabSize="normal"
-            app:layout_constraintBottom_toBottomOf="@+id/webview"
-            app:layout_constraintRight_toRightOf="@+id/webview"
-            app:srcCompat="@android:drawable/ic_delete"/>
+    <FrameLayout
+        android:id="@+id/work_activity_fragment_frame"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 
-        <Button
-            android:id="@+id/button_WorkActivity_8"
-            android:layout_width="200dp"
-            android:layout_height="@dimen/button_large_height"
-            android:layout_marginBottom="0dp"
-            android:layout_marginLeft="8dp"
-            android:layout_marginRight="8dp"
-            android:onClick="buttonFinishWebView"
-            android:text="@string/button_WorkActivity_8"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"/>
+    </LinearLayout>
 
-        <Button
-            android:id="@+id/button_WorkActivity_9"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Ok"
-            android:onClick="readInText"
-            android:layout_marginTop="0dp"
-            app:layout_constraintTop_toBottomOf="@+id/editText"
-            android:layout_marginRight="70dp"
-            app:layout_constraintRight_toRightOf="parent"/>
-
-        <android.support.design.widget.FloatingActionButton
-            android:id="@+id/button_WorkActivity_10"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginRight="8dp"
-            android:layout_marginTop="8dp"
-            android:layout_x="287dp"
-            android:layout_y="140dp"
-            android:clickable="true"
-            android:onClick="saveButton"
-            android:tint="?attr/editTextColor"
-            app:backgroundTint="?attr/colorButtonNormal"
-            app:fabSize="normal"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:srcCompat="@android:drawable/ic_menu_save"/>
-
-        <TextView
-            android:id="@+id/textView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="8dp"
-            android:layout_marginRight="8dp"
-            android:layout_marginTop="150dp"
-            android:text="Please fill in the PIN."
-            android:textAlignment="center"
-            android:textSize="24sp"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent"/>
-
-        <EditText
-            android:id="@+id/editText"
-            android:layout_width="330dp"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="0dp"
-            android:layout_marginRight="0dp"
-            android:layout_marginTop="50dp"
-            android:text=""
-            android:textAlignment="center"
-            android:textSize="18sp"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/textView"/>
-
-        <ListView
-            android:id="@+id/WorkActivityList"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:layout_marginBottom="8dp"
-            android:layout_marginTop="8dp"
-            app:layout_constraintBottom_toBottomOf="@+id/button_WorkActivity_8"
-            app:layout_constraintHorizontal_bias="0.0"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/work_activity_toolbar" />
-
-        <WebView
-            android:id="@+id/webview"
-            android:layout_width="fill_parent"
-            android:layout_height="fill_parent"
-            android:layout_marginBottom="-76dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="@+id/WorkActivityList" />
-
-    </android.support.constraint.ConstraintLayout>
 
     <android.support.design.widget.NavigationView
         android:id="@+id/nav_view"
