+++ app/src/main/res/values/strings.xml
@@ -82,4 +82,7 @@
     <string name="no_location_access">Error: No location access allowed!</string>
     <string name="no_external_storage_access">Error: No access to external storage allowed!</string>
 
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+
 </resources>
